requirements.txt
git+https://github.com/Rapptz/discord.py.git
discord.py[voice]
youtube_dl
dnspython